adjusted ordinal within the steps table to an integer instead of a float

+++ data/migrations/20191017183722_init.js
@@ -20,7 +20,7 @@ exports.up = function (knex) {
           .references('id')
           .inTable('recipes')
           .onDelete('cascade');
-        tbl.decimal('ordinal').notNullable();
+        tbl.integer('ordinal').notNullable();
         tbl.primary(['ordinal', 'recipe_id']);
         tbl.text('body').notNullable();
       })
